Fix small model weight loading (#184)
This pr fixes two issues:
- Small models that only have a single weight shard don't have a
`model.safetensors.index.json` file. Now we will fallback to checking
for `model.safetensors` if the index file doesn't exist
- Some models share the `lm_head` and `embed_tokens` weights and don't
have separate entries for both in the checkpoints. We now fallback to
using the `embed_tokens` value for the `lm_head` if we can't find an
`lm_head` entry.

---------

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>